Adjusted various components and pages to have correct paths and moved files into the correct folders
