+++ src/index.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import App from './App.js';
-import Profilescreen from './profilescreen.js'
+import Profilescreen from './pages/profilescreen.js'
+import Homepage from './pages/Homepage.js'
 import reportWebVitals from './reportWebVitals';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
-    <App />
+    <Homepage />
     <Profilescreen /> 
   </React.StrictMode>
 );
